👦🐑 Checkpoint ./commands/utility/sudo.js:26780611/2595

+++ commands/utility/sudo.js
@@ -1,20 +1,28 @@
 module.exports = {
-    name: 'sudo',
-    description: 'Makes a webhook to impersonate someone',
-    usage: 'sudo <user> <message>',
-    category: 'utility',
-    args: true,
-    cooldown: 5,
-    run: async (client, message, args) => {
-        message.delete()
-        let user = message.mentions.members.first() || message.guild.members.cache.get(args[0])
-        if(!user) return message.channel.send('Please provide a user!')
-        const webhook = await message.channel.createWebhook(user.displayName, {
-            avatar: user.user.displayAvatarURL(),
-            channel: message.channel.id
-        })
-        await webhook.send(args.slice(1).join(' ')).then(() => {
-            webhook.delete()
-        })
+  name: "sudo",
+  description: "Makes a webhook to impersonate someone",
+  usage: "sudo <user> <message>",
+  category: "utility",
+  args: true,
+  cooldown: 5,
+  run: async (client, message, args) => {
+  const m = client.user
+    if (!m.hasPermission("MANAGE_WEBHOOKS")) {
+      return message.channel.send(
+        `**${message.author.username}**, I do not have enough permission MANAGE_WEBHOOKS to use this command`
+      );
     }
-}
+     message.delete();
+    let user =
+      message.mentions.members.first() ||
+      message.guild.members.cache.get(args[0]);
+    if (!user) return message.channel.send("Please provide a user!");
+    const webhook = await message.channel.createWebhook(user.displayName, {
+      avatar: user.user.displayAvatarURL(),
+      channel: message.channel.id
+    });
+    await webhook.send(args.slice(1).join(" ")).then(() => {
+      webhook.delete();
+    });
+  }
+};
